+++ Hospital/src/main/java/Hospital/InpatientDepartment.java
@@ -4,26 +4,21 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
-public class InpatientDepartment {
+public class InpatientDepartment extends Department {
 
 	private int beds;
-	private ArrayList<Patient> patients;
-	private ArrayList<Staff> staff;
-	private String deptName;
+	protected HashMap<Integer, Patient> patients = new HashMap<Integer, Patient>();
 	
+
 	public InpatientDepartment(String deptName, int beds) {
-		this.deptName = deptName;
+		super(deptName);
 		this.beds = beds;
-		patients = new ArrayList<Patient>();
-		staff = new ArrayList<Staff>();
 	}
 	
-//	public Department(String deptName) {
-//		this.deptName = deptName;
-//		this.beds = 0;
-//		patients = new ArrayList<Patient>();
-//		staff = new ArrayList<Staff>();
-//	}
+	public InpatientDepartment(String deptName) {
+		super(deptName);
+		this.beds = 0;
+	}
 
 	/**
 	 *  returns the number of beds in a department
@@ -72,141 +67,5 @@ public class InpatientDepartment {
 		else {
 			this.beds -= beds;
 		}
-	
-	}
-	
-	/**
-	 * gets the name of a department
-	 * @return the name of the department
-	 */
-	public String getDeptName() {
-		return deptName;
-	}
-	
-	/**
-	 * sets the name of a department
-	 * @param deptName the name of the department to be set
-	 */
-	public void setDeptName(String deptName) {
-		this.deptName = deptName;
-	}
-	
-	/**
-	 * Adds a staff member to the hashmap of staffmembers in a department
-	 * @param s The staff member to be added
-	 */
-	protected void addStaff(Staff s) {
-		staff.add(s);
-	}
-	/**
-	 * Adds a patient to the hashmap of patients in a department
-	 * @param patient the patient to be added
-	 */
-	protected void addPatient(Patient patient) {
-		patients.add(patient);
-	}
-	/**
-	 * remove a patient from the arraylist of patientes in a department
-	 * @param patient patient to be removed
-	 */
-	protected void deletePatient(Patient patient) {
-		patients.remove(patient);
-	}
-	
-	/**
-	 * remove a staff member from the arraylist of staff in a department
-	 * @param staff staff to be removed
-	 */
-	protected void deleteStaff(Staff staff) {
-		this.staff.remove(staff);
-	}
-	
-	/**
-	 * Gives the patients in department 
-	 * @return string array of patients in department
-	 */
-	public String[] getPatients() {
-		ArrayList<String> patientslist = new ArrayList<String>();
-		for (Patient patient : patients) {
-			patientslist.add(patient.toString());
-		}
-		String[] patientListString = new String[patientslist.size()];
-		patientListString = patientslist.toArray(patientListString);
-		return patientListString;
-	}
-	
-	/**
-	 * Gives the staff members working in department
-	 * @return string array of staff members in department
-	 */
-	public String[] getStaff() {
-		ArrayList<String> stafflist = new ArrayList<String>();
-		for (Staff s : staff) {
-			stafflist.add(s.toString());
-		}
-		String[] staffListString = new String[stafflist.size()];
-		staffListString = stafflist.toArray(staffListString);
-		return staffListString;
-	}
-	
-	/**
-	 * Checks whether this department has a specified patient
-	 * @param patient A patient of type Patient
-	 * @return true if department has patient
-	 */
-	protected boolean containsPatient(Patient patient) {
-		return patients.contains(patient);
-	}
-	
-	/**
-	 * Checks whether this department has a patient with a matching serialnum
-	 * @param serialnum The patient's serialnum of type int
-	 * @return true if department has patient
-	 */
-	public boolean containsPatient(int serialnum) {
-		for (Patient patient : patients) {
-			if (patient.getSerialnum() == serialnum) {
-				return true;
-			}
-		}
-		return false;
-	}
-	
-	/**
-	 * Checks whether this department has a specified staff member
-	 * @param s A staff of type Staff
-	 * @return true if department has staff
-	 */
-	public boolean containsStaff(Staff s) {
-		return staff.contains(s);
-	}
-	
-	/**
-	 * Checks whether this department has a staff with a matching serialnum
-	 * @param serialnum The staff's serialnum of type int
-	 * @return true if department has staff
-	 */
-	public boolean containsStaff(int serialnum) {
-		for (Staff s : staff) {
-			if (s.getSerialnum() == serialnum) {
-				return true;
-			}
-		}
-		return false;
-	}
-	
-	/**
-	 * Gets the patient with a matching serialnum. Throws an IllegalArgumentException
-	 * returns null if there is no matching patient
-	 * @param serialnum
-	 * @return matching Patient p
-	 */
-	protected Patient getPatient(int serialnum) {
-		for (Patient patient : patients) {
-			if (patient.getSerialnum() == serialnum) {
-				return patient;
-			}
-		}
-		return null;
 	}
 }
